Source code/findcollab-master-Django-backEnd/requirements.txt
Django==1.11.14
django-taggit==0.22.1
pytz==2018.5
